+++ base/android/java/src/org/chromium/base/JniAndroid.java
@@ -53,8 +53,22 @@ public final class JniAndroid {
     public static class UncaughtExceptionException extends RuntimeException {
         public UncaughtExceptionException(String nativeStackTrace, Throwable uncaughtException) {
             super(
-                    "Native stack trace:" + System.lineSeparator() + nativeStackTrace,
+                    getTopFrameString(uncaughtException) + System.lineSeparator()
+                            + "Uncaught Java exception:" + System.lineSeparator()
+                            + Log.getStackTraceString(uncaughtException) + System.lineSeparator()
+                            + "Native stack trace:" + System.lineSeparator() + nativeStackTrace,
                     uncaughtException);
+            setStackTrace(uncaughtException.getStackTrace());
+        }
+
+        private static String getTopFrameString(Throwable uncaughtException) {
+            StackTraceElement[] frames = uncaughtException.getStackTrace();
+            if (frames != null && frames.length > 0) {
+                StackTraceElement top = frames[0];
+                String fileName = top.getFileName();
+                return (fileName != null ? fileName : "UnknownSource") + ":" + top.getLineNumber();
+            }
+            return "UnknownSource";
         }
     }