+++ starboard/android/shared/media_codec_bridge.h
@@ -134,7 +134,6 @@ class MediaCodecBridge {
       std::optional<int> tunnel_mode_audio_session_id,
       bool force_big_endian_hdr_metadata,
       int max_video_input_size,
-      bool enable_output_checker,
       bool skip_video_frames_over_60_fps);
 
   ~MediaCodecBridge();