+++ starboard/thread.h
@@ -32,53 +32,6 @@
 extern "C" {
 #endif
 
-// A spectrum of thread priorities. Platforms map them appropriately to their
-// own priority system. Note that scheduling is platform-specific, and what
-// these priorities mean, if they mean anything at all, is also
-// platform-specific.
-//
-// In particular, several of these priority values can map to the same priority
-// on a given platform. The only guarantee is that each lower priority should be
-// treated less-than-or-equal-to a higher priority.
-typedef enum SbThreadPriority {
-  // The lowest thread priority available on the current platform.
-  kSbThreadPriorityLowest,
-
-  // A lower-than-normal thread priority, if available on the current platform.
-  kSbThreadPriorityLow,
-
-  // Really, what is normal? You should spend time pondering that question more
-  // than you consider less-important things, but less than you think about
-  // more-important things.
-  kSbThreadPriorityNormal,
-
-  // A higher-than-normal thread priority, if available on the current platform.
-  kSbThreadPriorityHigh,
-
-  // The highest thread priority available on the current platform that isn't
-  // considered "real-time" or "time-critical," if those terms have any meaning
-  // on the current platform.
-  kSbThreadPriorityHighest,
-
-  // If the platform provides any kind of real-time or time-critical scheduling,
-  // this priority will request that treatment. Real-time scheduling generally
-  // means that the thread will have more consistency in scheduling than
-  // non-real-time scheduled threads, often by being more deterministic in how
-  // threads run in relation to each other. But exactly how being real-time
-  // affects the thread scheduling is platform-specific.
-  //
-  // For platforms where that is not offered, or otherwise not meaningful, this
-  // will just be the highest priority available in the platform's scheme, which
-  // may be the same as kSbThreadPriorityHighest.
-  kSbThreadPriorityRealTime,
-
-  // Well-defined constant value to mean "no priority."  This means to use the
-  // default priority assignment method of that platform. This may mean to
-  // inherit the priority of the spawning thread, or it may mean a specific
-  // default priority, or it may mean something else, depending on the platform.
-  kSbThreadNoPriority = INT_MIN,
-} SbThreadPriority;
-
 // An ID type that is unique per thread.
 typedef int32_t SbThreadId;
 
@@ -89,14 +42,6 @@ typedef int32_t SbThreadId;
 static inline bool SbThreadIsValidId(SbThreadId id) {
   return id != kSbThreadInvalidId;
 }
-
-// Returns whether the given thread priority is valid.
-static inline bool SbThreadIsValidPriority(SbThreadPriority priority) {
-  return priority != kSbThreadNoPriority;
-}
-
-// Returns the Thread ID of the currently executing thread.
-SB_EXPORT SbThreadId SbThreadGetId();
 
 #ifdef __cplusplus
 }  // extern "C"